requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Bootstrap4==4.0.2
python-dotenv==0.19.0
supabase==1.0.3
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
plotly==5.3.1
gunicorn==20.1.0
python-jose==3.3.0
passlib==1.7.4 